--- a/main/src/ca/uwaterloo/flix/language/phase/jvm/AsmOps.scala
+++ b/main/src/ca/uwaterloo/flix/language/phase/jvm/AsmOps.scala
@@ -302,7 +302,7 @@ object AsmOps {
     * For example, the class of `Tuple[Int32, Int32]` has the following `setField0` method:
     *
     * public final void setField0(int var) {
-    *   this.field0 = var;
+    * this.field0 = var;
     * }
     */
   def compileSetFieldMethod(visitor: ClassWriter, classType: JvmName, fieldName: String, methodName: String, fieldType: JvmType): Unit = {
@@ -640,7 +640,7 @@ object AsmOps {
   /**
     * Emits code to call a closure (not in tail position). fType is the type of the called closure. argsType is the type of its arguments, and resultType is the type of its result.
     */
-  def compileClosureApplication(visitor: MethodVisitor, fType: MonoType, argsTypes: List[MonoType], resultType: MonoType)(implicit root: Root, flix: Flix) = {
+  def compileClosureApplication(visitor: MethodVisitor, fType: MonoType, argsTypes: List[MonoType], resultType: MonoType, castFinalValue: Boolean = true)(implicit root: Root, flix: Flix): Unit = {
     // Type of the continuation interface
     val cont = JvmOps.getContinuationInterfaceType(fType)
     // Type of the function interface
@@ -679,6 +679,6 @@ object AsmOps {
     // Load IFO from local variable and invoke `getResult` on it
     visitor.visitVarInsn(ALOAD, 2)
     visitor.visitMethodInsn(INVOKEINTERFACE, cont.name.toInternalName, "getResult", AsmOps.getMethodDescriptor(Nil, JvmOps.getErasedJvmType(resultType)), true)
-    AsmOps.castIfNotPrim(visitor, JvmOps.getJvmType(resultType))
+    if (castFinalValue) AsmOps.castIfNotPrim(visitor, JvmOps.getJvmType(resultType))
   }
 }