Fixes #2328

Current benchmark:
```flix
use Benchmark.Benchmark;
use Benchmark.defBenchmark;

pub def bench1(): Benchmark = 
     defBenchmark("List", () -> { 
         List.range(1, 10_000) |>
         List.map(x -> x + 1) |>
         List.map(x -> x + 1) |>
         List.map(x -> x + 1)
    })

pub def bench2(): Benchmark = 
     defBenchmark("LazyList", () -> { 
         LazyList.range(1, 10_000) |>
         mapE(x -> x + 1) |>
         mapE(x -> x + 1) |>
         mapE(x -> x + 1 as & Impure)
    } as & Pure)

pub def mapE(f: a -> b & ef, l: LazyList[a]): LazyList[b] & Impure =
    let r = MutList.new();
    mapERec(f, l, r);
    MutList.foldRight((x, xs) -> ECons(x, xs), ENil, r)

def mapERec(f: a -> b & ef, l: LazyList[a], r: MutList[b]): Unit & Impure =
    match l {
        case ENil         => ()
        case ECons(x, xs) => 
            MutList.push!(f(x), r);
            mapERec(f, xs, r)
        case LCons(x, xs) => 
            MutList.push!(f(x), r);
            mapERec(f, force xs, r)
        case LList(xs)    => mapERec(f, force xs, r)
    }
    
def main(_args: Array[String]) : Int32 & Impure = 
    let bs = [bench1(), bench2()];
    Benchmark.runWithBudget(bs, 240_000_000_000i64);
    0
    

```
Result:
```
Benchmark                                         Iterations          Time (avg)
List                                                    6829          4836 us/op
LazyList                                                1638         18310 us/op

Total time elapsed: 63 seconds. Time budget was: 240 seconds.
```