@magnus-madsen 2021-09-17
How does it compare before vs. after?