@JonathanStarup 2021-09-17
There is also a question about exceptions. Right now exceptions stop lock unlocking (which ill fix here), but an exception would still break the value initialization part, such that the next need for the lazy value re-evaluates the thunk, but I suppose it should only rethrow the exception but not reevaluate other sideeffect. My current idea is:
```java
package ca.uwaterloo.flix;

class LazyInt<A> {
    public A value;
    public boolean init;
    private Cont<A> thunk;

    public LazyInt(Cont<A> thunk) {
        this.thunk = thunk;
        this.init = false;
    }

    public A force() {
        synchronized (this) {
            if (init) return value; // avoid double computation when threads are racing to initialize
            value = thunk.unwind();
            init = true;
        }
        thunk = null; // locking not needed as init and value are setup which will handle the other use cases
        return value;
    }

    // this should be generated inline, not called
    public A getValue() {
        if (init) return value;
        else return force();
    }

}

interface Cont<A> {
    A unwind();
}

```