@JonathanStarup 2021-09-17
Lazy Obj
```java
import Cont$Obj;

public final class Lazy$Obj {
    public Object expression;
    public Object value;

    public final synchronized Object force(Context var1) {
        if (this.expression == null) {
            return this.value;
        } else {
            var1.continuation = this.expression;

            Obj var2;
            do {
                Object var10001 = var1.continuation;
                var1.continuation = null;
                (var2 = (Cont$Obj)var10001).invoke(var1);
            } while(var1.continuation != null);

            this.value = var2.getResult();
            this.expression = null;
            return this.value;
        }
    }

    public Lazy$Obj(Object var1) {
        this.expression = var1;
    }
}
```
Results:
```
...
```