@JonathanStarup 2021-09-17
Without the synchronzied method, using the old flawed Monitorenter/exit:
```
Benchmark                                         Iterations          Time (avg)
List                                                   15530          4571 us/op
LazyList                                                3554         15429 us/op

Total time elapsed: 125 seconds. Time budget was: 300 seconds.
```
```
Benchmark                                         Iterations          Time (avg)
List                                                    6228          4394 us/op
LazyList                                                1457         18106 us/op

Total time elapsed: 53 seconds. Time budget was: 150 seconds.
```