@JonathanStarup 2021-09-17
Time budget | Master | Reduced locking | Reduced locking + synchronized force
--- | --- | --- | ---
150 | x3.81 | x4.12 | x3.33
300 | x4.39 | x3.38 | x3.11
600 | x3.24 | x3.39 | x2.89