@magnus-madsen 2021-09-17
By the way, I guess the `expression` can still be private.