@JonathanStarup 2021-09-17
The tests here are a bit volatile, but it seems a faster yes.

expression is used in compileExp in genExpression to check lazy.expression == null.